Address second round of isl_utility review feedback
- Move isl_utility.py into visitor_registry/ (its only consumers) instead
  of a new top-level utilities/ package; drop the now-empty utilities/
  directories (src and tests). Test file follows to tests/analysis/,
  alongside test_relation_build.py.
- dim_range now builds an isl value expression for the dim (native +/-/*
  /floor/mod/max/min, arbitrarily nested) and reads the range back from
  isl's own min_val/max_val, instead of hand-rolled interval arithmetic.
  isl cannot represent a product of two non-constant terms at all (not
  affine), so that one case still computes via 4-corner interval math,
  recursing back through dim_range for its operands' own ranges. A
  symbolic (non-constant) floordiv/mod divisor is equally unrepresentable
  in isl and isn't exercised by any acceptance case, so it now raises
  NotImplementedError instead of a hand-rolled fallback.
- Drop the ParamMap/DomainBuild aliases; to_domain/to_dim take and return
  plain dict/tuple.
- Trim comments throughout to what the code doesn't already say on its
  own, and remove every nncase reference from source and test comments.

Re-verified: repro_g02_g03.py 5/5, repro_g02_g03b.py 3/3, pytest tests/ -q
850 passed, spec/entropy lints and ruff clean, no module-level mutable
dict in the diff vs main.
