+++ src/tilefoundry/ir/hir/nn/matmul.py
@@ -16,12 +16,12 @@ from tilefoundry.ir.types.shard.shard_layout import (
     Split,
     layout_axis_to_tensor_axis,
 )
-from tilefoundry.utilities.isl_utility import to_domain
 from tilefoundry.visitor_registry.access_relation import (
     AccessRelationResult,
     build_relation,
     register_type_relation,
 )
+from tilefoundry.visitor_registry.isl_utility import to_domain
 from tilefoundry.visitor_registry.relation_build import shape_from_relation
 from tilefoundry.visitor_registry.shard_propagate import derive_output_shard_layout
