--- src/tilefoundry/utilities/__init__.py
@@ -1,2 +1,0 @@
-"""``tilefoundry.utilities`` — standalone helpers shared across subsystems."""
-from __future__ import annotations
