+++ src/tilefoundry/visitor_registry/relation_build.py
@@ -1,34 +1,26 @@
-"""Forward access-relation construction helpers (input-type driven).
-
-Builds the bounded iteration ``domain`` (an ``isl.set``) for an op from its
-iteration extents, and recovers an output shape from a built relation. Both
-directions are a thin consumer of ``tilefoundry.utilities.isl_utility`` (the
-``ShapeDim`` <-> isl bridge) — this module owns no ``ShapeDim`` <-> isl
-translation of its own.
-"""
+"""Forward access-relation construction helpers (input-type driven)."""
 from __future__ import annotations
 
 import isl
 
-from tilefoundry.utilities.isl_utility import to_dim, to_domain
+from tilefoundry.visitor_registry.isl_utility import to_dim, to_domain
 
 
 def build_domain(extents: tuple) -> "isl.set":
-    """Bounded iteration domain ``{ [d0, ..., dn] : 0 <= di < extent_i }``
-    (see ``isl_utility.to_domain``). For a relation whose output shape will
-    be recovered via ``shape_from_relation``, call ``to_domain`` directly
-    instead, so its ``param_map`` travels alongside the relation."""
-    return to_domain(extents).domain
+    """Bounded iteration domain for *extents* (see ``isl_utility.to_domain``).
+    A caller that also needs the output shape back (``shape_from_relation``)
+    should call ``to_domain`` directly instead, to keep its param map
+    alongside the relation."""
+    domain, _ = to_domain(extents)
+    return domain
 
 
 def shape_from_relation(relation) -> tuple:
     """Derive the output shape from the relation's output map + bounded domain.
 
-    Each output map result axis is a pure projection of a domain dim — its
-    extent (``domain.dim_max(d) + 1``) is recovered as a ``ShapeDim`` via
-    ``isl_utility.to_dim`` and ``relation.param_map`` — or a constant (a
-    size-1 output axis). A non-projection / non-constant result axis, or an
-    extent that does not resolve to a ShapeDim, fails closed.
+    Each output map result axis is a pure projection of a domain dim (its
+    extent is recovered via ``isl_utility.to_dim`` and ``relation.param_map``)
+    or a constant (a size-1 output axis). Anything else fails closed.
     """
     domain = relation.domain
     output_map = relation.maps[-1]
@@ -44,7 +36,7 @@ def shape_from_relation(relation) -> tuple:
             if int(aff.get_coefficient_val(isl.dim_type.IN, j).num_si()) != 0
         ]
         if not used:
-            shape.append(1)  # constant result: a size-1 output axis
+            shape.append(1)
         elif len(used) == 1 and used[0][1] == 1:
             extent = domain.dim_max(used[0][0]).add_constant(1)
             shape.append(to_dim(extent, relation.param_map))
